+++ Assets/Libraries/Entitas/Entitas/CodeGenerator/Attributes/CustomPrefixAttribute.cs
@@ -2,7 +2,7 @@
 
 namespace Entitas.CodeGenerator {
 
-    [AttributeUsage(AttributeTargets.Class)]
+    [AttributeUsage(AttributeTargets.Class | AttributeTargets.Interface | AttributeTargets.Struct)]
     public class CustomPrefixAttribute : Attribute {
         public readonly string prefix;