+++ Assets/Libraries/Entitas/Entitas/Unity/VisualDebugging/Editor/EntitasHierarchyIcon.cs
@@ -9,7 +9,7 @@ namespace Entitas.Unity.VisualDebugging {
         static Texture2D poolHierarchyIcon {
             get {
                 if (_poolHierarchyIcon == null) {
-                    _poolHierarchyIcon = EntitasEditorLayout.LoadTexture("l:PoolHierarchyIcon");
+                    _poolHierarchyIcon = EntitasEditorLayout.LoadTexture("l:EntitasPoolHierarchyIcon");
                 }
                 return _poolHierarchyIcon;
             }
@@ -18,7 +18,7 @@ namespace Entitas.Unity.VisualDebugging {
         static Texture2D poolErrorHierarchyIcon {
             get {
                 if (_poolErrorHierarchyIcon == null) {
-                    _poolErrorHierarchyIcon = EntitasEditorLayout.LoadTexture("l:PoolErrorHierarchyIcon");
+                    _poolErrorHierarchyIcon = EntitasEditorLayout.LoadTexture("l:EntitasPoolErrorHierarchyIcon");
                 }
                 return _poolErrorHierarchyIcon;
             }
@@ -26,17 +26,26 @@ namespace Entitas.Unity.VisualDebugging {
 
         static Texture2D entityHierarchyIcon {
             get {
-                if (_entityhierarchyIcon == null) {
-                    _entityhierarchyIcon = EntitasEditorLayout.LoadTexture("l:EntityHierarchyIcon");
+                if (_entityHierarchyIcon == null) {
+                    _entityHierarchyIcon = EntitasEditorLayout.LoadTexture("l:EntitasEntityHierarchyIcon");
                 }
-                return _entityhierarchyIcon;
+                return _entityHierarchyIcon;
+            }
+        }
+
+        static Texture2D entityErrorHierarchyIcon {
+            get {
+                if (_entityErrorHierarchyIcon == null) {
+                    _entityErrorHierarchyIcon = EntitasEditorLayout.LoadTexture("l:EntitasEntityErrorHierarchyIcon");
+                }
+                return _entityErrorHierarchyIcon;
             }
         }
 
         static Texture2D systemsHierarchyIcon {
             get {
                 if (_systemsHierarchyIcon == null) {
-                    _systemsHierarchyIcon = EntitasEditorLayout.LoadTexture("l:SystemsHierarchyIcon");
+                    _systemsHierarchyIcon = EntitasEditorLayout.LoadTexture("l:EntitasSystemsHierarchyIcon");
                 }
                 return _systemsHierarchyIcon;
             }
@@ -44,7 +53,8 @@ namespace Entitas.Unity.VisualDebugging {
 
         static Texture2D _poolHierarchyIcon;
         static Texture2D _poolErrorHierarchyIcon;
-        static Texture2D _entityhierarchyIcon;
+        static Texture2D _entityHierarchyIcon;
+        static Texture2D _entityErrorHierarchyIcon;
         static Texture2D _systemsHierarchyIcon;
 
         static EntitasHierarchyIcon() {
@@ -59,15 +69,22 @@ namespace Entitas.Unity.VisualDebugging {
                 var rect = new Rect(selectionRect.x + selectionRect.width - iconOffset, selectionRect.y, iconSize, iconSize);
 
                 var poolObserver = gameObject.GetComponent<PoolObserverBehaviour>();
+                var entityBehaviour = gameObject.GetComponent<EntityBehaviour>();
+                var debugSystemsBehaviour = gameObject.GetComponent<DebugSystemsBehaviour>();
+
                 if (poolObserver != null) {
                     if (poolObserver.poolObserver.pool.retainedEntitiesCount != 0) {
                         GUI.DrawTexture(rect, poolErrorHierarchyIcon);
                     } else {
                         GUI.DrawTexture(rect, poolHierarchyIcon);
                     }
-                } else if (gameObject.GetComponent<EntityBehaviour>() != null) {
-                    GUI.DrawTexture(rect, entityHierarchyIcon);
-                } else if (gameObject.GetComponent<DebugSystemsBehaviour>() != null) {
+                } else if (entityBehaviour != null) {
+                    if (entityBehaviour.entity.isEnabled) {
+                        GUI.DrawTexture(rect, entityHierarchyIcon);
+                    } else {
+                        GUI.DrawTexture(rect, entityErrorHierarchyIcon);
+                    }
+                } else if (debugSystemsBehaviour != null) {
                     GUI.DrawTexture(rect, systemsHierarchyIcon);
                 }
             }