+++ Assets/Libraries/Entitas/RELEASE_NOTES.md
@@ -1,3 +1,76 @@
+# 0.31.2
+
+##### Entitas.CodeGenerator
+- All attributes can now be used for classes, interfaces and structs
+
+
+# 0.31.1
+
+##### Entitas.CodeGenerator
+- Improved component generation for classes and interfaces and added support for default pool [Pool]
+- Added support to CustomComponentNameAttribute to generate multiple components with different names for one class or interface
+
+```csharp
+// This will automatically generate PositionComponent and VelocityComponent for you
+[Pool, CustomComponentName("Position", "Velocity")]
+public struct IntVector2 {
+  public int x;
+  public int y;
+}
+```
+
+- Added support for generating components for structs
+- Not generating obsolete pool attributes for generated classes
+
+
+# 0.31.0
+
+##### General
+- Removed obsolete code
+
+##### Entitas.CodeGenerator
+- Generating components for attributed classes and interfaces
+
+```csharp
+// will automatically generate SomeClassComponent for you
+[Core]
+public class SomeClass {
+    public string name;
+
+    public SomeClass(string name) {
+        this.name = name;
+    }
+}
+```
+
+- Added support to add empty PoolAttribute to assign component to default pool
+
+```csharp
+// using [Pool] will also add this component to Pools.pool
+[Core, Pool]
+public class SomeComponent : IComponent {
+}
+```
+
+##### Entitas.Unity.VisualDebugging
+- Added IComponentDrawer which can draw the whole component
+
+##### Entitas.Unity.VisualDebugging
+- Added EntitasEntityErrorHierarchyIcon to indicate retained entities in the hierarchy
+- Added CharTypeDrawer
+- Fixed components not updating in the inspector (#107)
+- Improved SystemsMonitor and added average line
+
+![Entitas-SystemsMonitor](https://cloud.githubusercontent.com/assets/233700/15198441/a515d764-17d7-11e6-965c-83c027fa89f7.png)
+
+##### Entitas.Unity.Serialization.Blueprints
+- Fixed finding all BinaryBlueprints even when not loaded
+- Correctly saving Blueprints when setting all BinaryBlueprints
+- Added BlueprintsNotFoundException
+- BinaryBlueprintInspector creates new pools instead of using one of Pools.allPools
+- Fixed pool not shown when entering play-mode while a blueprint was selected in the project view
+- Not caching blueprints when UNITY_EDITOR to enable live edit
+
 # 0.30.3
 
 ##### Entitas.CodeGenerator